requirements.txt
beautifulsoup4==4.11.1
cssutils==2.5.1
httpx==0.23.0
lxml==4.9.1
js2py==0.71
rich==12.5.1
m3u8==3.2.0
alive-progress==2.4.1
filetype==1.1.0
pyperclip==1.8.2
jinja2==3.1.2
arrow==1.2.2
websocket-client==1.4.0
protobuf==3.20.1
pycryptodome==3.15.0
